fix: use roles rather than scopes for jupyterhub clients
We are currently using scopes to indicate permissions within auth tokens for Jupyterhub. For the sake of simplicity, we used common scopes across each hub deployment. However, this introduced an unintended side-effect in that if a user had access to a scope via a Client->Role->Group relationship, then they would have access to that Scope for any client. This meant that we were unable to have separate permissions across clients. By instead utilizing Roles to identify a permission, we are able to employ a the `oidc-usermodel-client-role-mapper` to inject only roles matching the client.

This alters the auth token from:

```json
{
  # ...
  "scope": "openid profile jupyterhub:cpu:xs jupyterhub:cpu:s jupyterhub:cpu:m jupyterhub:cpu:l jupyterhub:cpu:xl jupyterhub:cpu:xxl jupyterhub:cpu:xxxl",
}
```

to:

```json
{
  # ...
  "scope": "openid profile",
  "roles": [
    "CPU:XXXL",
    "CPU:S",
    "CPU:XXL",
    "CPU:XL",
    "CPU:M",
    "CPU:XS",
    "CPU:L"
  ]
}
```